Restructure in paths
