+++ pkg/paths/use.go
@@ -1,7 +1,6 @@
 package paths
 
 import (
-	"os"
 	"path/filepath"
 
 	"github.com/tfversion/tfversion/pkg/helpers"
@@ -20,19 +19,4 @@ func GetUseLocation() string {
 // GetActiveBinaryLocation returns the path to the currently active Terraform binary.
 func GetActiveBinaryLocation() string {
 	return filepath.Join(GetUseLocation(), TerraformBinaryName)
-}
-
-// GetUseVersion returns the currently active Terraform version.
-func GetUseVersion() string {
-	binaryPath := GetActiveBinaryLocation()
-	_, err := os.Lstat(binaryPath)
-	if err != nil {
-		helpers.ExitWithError("no current terraform version found", err)
-	}
-	realPath, err := filepath.EvalSymlinks(binaryPath)
-	if err != nil {
-		helpers.ExitWithError("resolving symlink", err)
-	}
-	_, version := filepath.Split(filepath.Dir(realPath))
-	return version
 }
